+++ NEWS.md
@@ -1,5 +1,8 @@
 # ellmer (development version)
 
+* New `interpolate_package()` to make it easier to interpolate from prompts
+  stored in the `inst/prompts` inside a package (#164).
+
 * `chat_claude()`, `chat_openai()`, and `chat_gemini()` now have a `params` 
   argument that allows you to specify common model paramaters (like `seed`
   and `temperature`). Support for other models will grow as you request it